Remove bootstrap address get header support (#10)
Bootstrap address doesn't really have a header. Its header is formed
implicitly by its structure - that's the intended design by IOHK.
However it does have an id, which emerges from its CBOR structure.
I left the ID in the code in case it might be needed later, However
an Error will be raised if someone tries to retrieve the whole header
for bootstrap address. I raise ValueError instead of wire.DataError,
because if get header is called with bootstrap address, it's probably
a developer error.