Redo keychain (seed.py) after rebasing
We need to have both the Byron and Shelley path namespace available
at the same time. I've done this once already, but the keychains
were redesigned by Trezor in their master, so I had to do it again
for the new keychain.

I have separated the changes, so that they can be properly
reviewed because they are very important.